fix(guard): close native policy review findings
