+++ rust/crates/guard-rule-contract/src/lib.rs
@@ -6,7 +6,7 @@ use sha2::{Digest, Sha256};
 pub const RULE_CONTRACT_SCHEMA: &str = "hol-guard-native-rule-contract.v2";
 const RULE_CONTRACT_DOMAIN: &[u8] = b"hol-guard-native-rule-contract.v2\0";
 
-const COMPONENTS: [(&str, &[u8]); 10] = [
+const COMPONENTS: [(&str, &[u8]); 15] = [
     (
         "guard-rules",
         include_bytes!("../../guard-rules/src/lib.rs"),
@@ -46,6 +46,26 @@ const COMPONENTS: [(&str, &[u8]); 10] = [
     (
         "guard-runtime-policy-enforcement",
         include_bytes!("../../guard-runtime/src/policy_enforcement.rs"),
+    ),
+    (
+        "guard-runtime-policy-enforcement-facts",
+        include_bytes!("../../guard-runtime/src/policy_enforcement_facts.rs"),
+    ),
+    (
+        "guard-runtime-policy-enforcement-policy",
+        include_bytes!("../../guard-runtime/src/policy_enforcement_policy.rs"),
+    ),
+    (
+        "guard-policy-snapshot",
+        include_bytes!("../../guard-policy-snapshot/src/lib.rs"),
+    ),
+    (
+        "guard-policy-snapshot-canonical",
+        include_bytes!("../../guard-policy-snapshot/src/policy_snapshot_canonical.rs"),
+    ),
+    (
+        "guard-policy-snapshot-crypto",
+        include_bytes!("../../guard-policy-snapshot/src/policy_snapshot_crypto.rs"),
     ),
 ];
 
@@ -122,6 +142,11 @@ mod tests {
                 "guard-command-pretool-result",
                 "guard-command-pretool-extract",
                 "guard-runtime-policy-enforcement",
+                "guard-runtime-policy-enforcement-facts",
+                "guard-runtime-policy-enforcement-policy",
+                "guard-policy-snapshot",
+                "guard-policy-snapshot-canonical",
+                "guard-policy-snapshot-crypto",
             ]
         );
         assert!(first
