+++ scripts/bench_guard_native_full_path.py
@@ -13,12 +13,13 @@ import os
 import statistics
 import tempfile
 import time
+from collections.abc import Mapping
 from pathlib import Path
 
+from ci.native_runtime.native_policy_test_support import native_policy_snapshot
 from codex_plugin_scanner.guard.codex_hook_launch_runtime import run_isolated_hook_process
 from codex_plugin_scanner.guard.daemon.hook_process_runner import HookProcessRunner
 from codex_plugin_scanner.guard.native_runtime import (
-    native_runtime_status,
     review_post_tool_native,
 )
 from codex_plugin_scanner.guard.native_runtime_resident import close_resident_native_runtimes
@@ -130,6 +131,7 @@ def _bench_native_warm(
     workspace: Path,
     guard_home: Path,
     iterations: int,
+    policy_snapshot: Mapping[str, object],
 ) -> list[float]:
     values: list[float] = []
     for index in range(iterations):
@@ -142,7 +144,7 @@ def _bench_native_warm(
             }
         )
         started = time.perf_counter()
-        response = review_post_tool_native(request, observe_mode=False)
+        response = review_post_tool_native(request, observe_mode=False, policy_snapshot=policy_snapshot)
         elapsed = (time.perf_counter() - started) * 1_000.0
         if response is None or response.decision != "allow":
             raise RuntimeError("Native resident runtime did not return the expected allow decision")
@@ -248,26 +250,29 @@ def main() -> int:
                 iterations=2,
             )
             close_resident_native_runtimes()
-            first_native_started = time.perf_counter()
-            first_native = review_post_tool_native(
-                _native_request(workspace=workspace, guard_home=guard_home),
-                observe_mode=False,
-            )
-            native_readiness_ms = (time.perf_counter() - first_native_started) * 1_000.0
-            if first_native is None or first_native.decision != "allow":
-                raise RuntimeError("Native resident runtime readiness probe failed")
+            with native_policy_snapshot(guard_home) as snapshot:
+                first_native_started = time.perf_counter()
+                first_native = review_post_tool_native(
+                    _native_request(workspace=workspace, guard_home=guard_home),
+                    observe_mode=False,
+                    policy_snapshot=snapshot,
+                )
+                native_readiness_ms = (time.perf_counter() - first_native_started) * 1_000.0
+                if first_native is None or first_native.decision != "allow":
+                    raise RuntimeError("Native resident runtime readiness probe failed")
 
-            python_warm = _bench_python_warm(
-                runner=python_runner,
-                workspace=workspace,
-                guard_home=guard_home,
-                iterations=args.warm_iterations,
-            )
-            native_warm = _bench_native_warm(
-                workspace=workspace,
-                guard_home=guard_home,
-                iterations=args.warm_iterations,
-            )
+                python_warm = _bench_python_warm(
+                    runner=python_runner,
+                    workspace=workspace,
+                    guard_home=guard_home,
+                    iterations=args.warm_iterations,
+                )
+                native_warm = _bench_native_warm(
+                    workspace=workspace,
+                    guard_home=guard_home,
+                    iterations=args.warm_iterations,
+                    policy_snapshot=snapshot,
+                )
         finally:
             python_runner.close()
             close_resident_native_runtimes()
